--- app/src/languages.js
@@ -1,12 +1,0 @@
-const languages = {
-  en: {
-    name: 'English',
-    common: require('../build/locales/en/common.json'),
-  },
-  de: {
-    name: 'Deutsch',
-    common: require('../build/locales/de/common.json'),
-  },
-};
-
-export default languages;
