+++ i18n/languages.js
@@ -1,0 +1,13 @@
+const languages = {
+  en: {
+    name: 'English',
+    common: require('./locales/en/common.json'),
+    flag: require('./flags/en.png'),
+  },
+  de: {
+    name: 'Deutsch',
+    common: require('./locales/de/common.json'),
+    flag: require('./flags/de.png'),
+  },
+};
+export default languages;
