+++ package.json
@@ -18,7 +18,7 @@
     "dist:win": "build --win --ia32 --x64",
     "dist:mac": "build --mac",
     "dist:linux": "build --linux --ia32 --x64 --armv7l",
-    "copy-files": "cpx \"./src/{index.html,assets/**/*,locales/**/*}\" ./app/build/",
+    "copy-files": "cpx \"./src/{index.html,assets/**/*}\" ./app/build/",
     "clean-build": "rm -rf app/build",
     "clean-dist": "rm -rf dist",
     "eslint": "eslint ./src/ ./app/src/ ./app/config/ ./features/",
