--- src/constants/languages.js
@@ -1,14 +1,0 @@
-const languages = {
-  en: {
-    name: 'English',
-    common: require('../locales/en/common.json'),
-    flag: require('../assets/images/flags/en.png'),
-  },
-  de: {
-    name: 'Deutsch',
-    common: require('../locales/de/common.json'),
-    flag: require('../assets/images/flags/de.png'),
-  },
-};
-
-export default languages;
