+++ src/i18n.js
@@ -1,43 +1,5 @@
-import i18n from 'i18next';
-import languages from './constants/languages';
+import setup from '../i18n/i18n-setup';
 
-const resources = Object.keys(languages).reduce((accumulator, key) => {
-  accumulator[key] = {
-    common: languages[key].common,
-  };
-  return accumulator;
-}, {});
-
-i18n
-  .init({
-    fallbackLng: 'en',
-    resources,
-    react: {
-      // wait: true, // globally set to wait for loaded translations in translate hoc
-      // exposeNamespace: true // exposes namespace on data-i18next-options to be used in eg.
-    },
-
-    // have a common namespace used around the full app
-    ns: ['common'],
-    defaultNS: 'common',
-
-    keySeparator: '>',
-    nsSeparator: '|',
-
-    debug: false,
-
-    // cache: {
-    //   enabled: true,
-    // },
-
-    interpolation: {
-      escapeValue: false, // not needed for react!!
-      formatSeparator: ',',
-      format: (value, format) => {
-        if (format === 'uppercase') return value.toUpperCase();
-        return value;
-      },
-    },
-  });
+const i18n = setup();
 
 export default i18n;
